The JWT authentication service has been moved to the Core layer, and the Serilog and JWT configurations are now loaded from a shared appsettings.shared.json file located in the Core layer.

+++ src/BuildingBlocks/Services/Identity/Identity.Application/Concretes/JwtService.cs
@@ -1,16 +1,16 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
+using Core.Configuration.Settings;
 using Identity.Application.Contracts;
 using Identity.Application.Features.Shared.DTOs;
 using Identity.Application.Features.Users.Queries.Login;
 using Identity.Domain.Entities;
 using Identity.Persistence.Context;
-using Identity.Persistence.Options;
 using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
-using Microsoft.EntityFrameworkCore;
 
 namespace Identity.Application.Concretes;
 
@@ -37,12 +37,12 @@ public class JwtService(
         claims.AddRange(roles.Select(x => new Claim(ClaimTypes.Role, x)));
 
         var jwt = new JwtSecurityToken(
-            issuer: tokenOptions.Value.Issuer,
-            audience: tokenOptions.Value.Audience,
-            claims: claims,
-            notBefore: dateTimeNow,
-            expires: dateTimeNow.AddMinutes(tokenOptions.Value.ExpireInMinute),
-            signingCredentials: new SigningCredentials(key, SecurityAlgorithms.HmacSha256)
+            tokenOptions.Value.Issuer,
+            tokenOptions.Value.Audience,
+            claims,
+            dateTimeNow,
+            dateTimeNow.AddMinutes(tokenOptions.Value.ExpireInMinute),
+            new SigningCredentials(key, SecurityAlgorithms.HmacSha256)
         );
 
         var accessToken = new JwtSecurityTokenHandler().WriteToken(jwt);
@@ -50,10 +50,7 @@ public class JwtService(
         var oldTokens = await context.RefreshTokens
             .Where(r => r.UserId == user.Id && r.Expires < DateTime.UtcNow)
             .ToListAsync();
-        if (oldTokens.Count > 0)
-        {
-            context.RefreshTokens.RemoveRange(oldTokens);
-        }
+        if (oldTokens.Count > 0) context.RefreshTokens.RemoveRange(oldTokens);
 
         var refreshToken = Convert.ToBase64String(Guid.NewGuid().ToByteArray())
             .Replace("+", "").Replace("/", "").Replace("=", "");
